[services] Extract hostListFromChunks function into services/batch module
